+++ community-service/src/main/java/project/communityservice/domain/board/post/dto/request/BoardCreateRequest.java
@@ -1,4 +1,4 @@
-package project.communityservice.domain.board.dto.request;
+package project.communityservice.domain.board.post.dto.request;
 
 public record BoardCreateRequest(
         String title,
